Reactive Contact Form
The contact form is a vue app, it is reactive.
The first call is fetching data from the database to populate the volunteer interest options on the form.

The second call connects to a route that sends the form to the backend framework, which handles updating the data in the database.